+++ inc/System/System.h
@@ -26,7 +26,7 @@ void    *Sys_LoadLibrary(const char *path, const char *sym, void **handle);
 void    Sys_FreeLibrary(void *handle);
 void    *Sys_GetProcAddress(void *handle, const char *sym);
 
-unsigned    Sys_Milliseconds(void);
+int64_t    Sys_Milliseconds(void);
 void    Sys_Sleep(int msec);
 qboolean Sys_IsDir(const char *path);
 qboolean Sys_IsFile(const char *path);
